Refatorando alguns erros
